Signed-Backend/SignedBackend/requirements.txt
anyio==4.11.0
asgiref==3.9.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.7
firebase-admin==6.4.0
gcloud==0.18.3
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jwcrypto==1.5.6
msgpack==1.1.1
oauth2client==4.1.3
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
Pyrebase4==4.7.1
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.1
python-jwt==4.1.0
pytz==2025.2
PyYAML==6.0.3
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
whitenoise==6.6.0
